Move search bar and data grid into common/hooks folder
