+++ plugins/main/public/components/common/hooks/data_grid/use_data_grid.ts
@@ -3,6 +3,7 @@ import React, { useEffect, useMemo, useState, Fragment } from "react";
 import { SearchResponse } from "@opensearch-project/opensearch/api/types";
 // ToDo: check how create this methods
 import { parseData, getFieldFormatted, parseColumns } from './data_grid_service';
+import { IndexPattern } from '../../../../../../../src/plugins/data/common';
 
 const MAX_ENTRIES_PER_QUERY = 10000;
 
@@ -22,7 +23,7 @@ type tDataGridProps = {
 export const useDataGrid = (props: tDataGridProps): EuiDataGridProps => {
     const { indexPattern, DocViewInspectButton, results, defaultColumns } = props;
     /** Columns **/
-    const [columns, setColumns] = useState<EuiDataGridColumn[]>(defaultColumns);
+    const [columns, setColumns] = useState<tDataGridColumn[]>(defaultColumns);
     const [columnVisibility, setVisibility] = useState(() =>
         columns.map(({ id }) => id)
     );
