+++ plugins/main/public/components/common/hooks/search_bar/index.ts
@@ -1,0 +1,2 @@
+export * from './use_search_bar_service';
+export * from './use_search_bar';
