+++ plugins/main/public/components/common/wazuh_discover/wz_discover.tsx
@@ -23,15 +23,15 @@ import { SearchResponse } from '../../../../../../src/core/server';
 import DocViewer from '../doc_viewer/doc_viewer';
 import { DiscoverNoResults } from '../../overview/vulnerabilities/common/components/no_results';
 import { LoadingSpinner } from '../../overview/vulnerabilities/common/components/loading_spinner';
-import { useDataGrid, tDataGridColumn } from '../data_grid';
+import { useDataGrid, tDataGridColumn, exportSearchToCSV } from '../hooks/data_grid';
 import { useDocViewer } from '../doc_viewer/use_doc_viewer';
 //import './inventory.scss';
-import { search, exportSearchToCSV } from '../../overview/vulnerabilities/dashboards/inventory/inventory_service';
 import { ErrorHandler, ErrorFactory, HttpError } from '../../../react-services/error-management';
 import { withErrorBoundary } from '../hocs';
 import { HitsCounter } from '../../../kibana-integrations/discover/application/components/hits_counter';
 import { formatNumWithCommas } from '../../../kibana-integrations/discover/application/helpers';
-import useSearchBar from '../search_bar/use_search_bar';
+import useSearchBar from '../hooks/search_bar/use_search_bar';
+import { search } from '../hooks/search_bar';
 import { getPlugins } from '../../../kibana-services';
 import { ViewMode } from '../../../../../../src/plugins/embeddable/public';
 import { getDiscoverPanels } from './config/chart';
@@ -42,9 +42,6 @@ const DashboardByRenderer = getPlugins().dashboard.DashboardContainerByValueRend
  * - add possibility to customize column render
  * - add save query feature
  */
-
-
-
 
 export const MAX_ENTRIES_PER_QUERY = 10000;
